Suppressed warnings for 3rd party sources & expert discovery plugin; fixed almost all gcc warnings in our code along with few bugs
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@658 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Algorithm/src/util_algorithm/BinaryFindOpenCL.h
@@ -46,7 +46,7 @@ private:
     const NumberType* numbers;
     const int numbersSize;
     const NumberType* findNumbers;
-    const int findNumbersSize;
+    const size_t findNumbersSize;
     const NumberType filter;
 
     cl_event clEvent1;
